store hamocc ann_ts files

+++ packages/HAMOCC_DIAG/code/compute_ann_time_series.sh
@@ -4,8 +4,9 @@ script_start=`date +%s`
 #
 # HAMOCC DIAGNOSTICS package: compute_ann_time_series.sh
 # PURPOSE: computes annual time series from annual or monthly history files
-# Johan Liakka, NERSC, johan.liakka@nersc.no
-# Last update Dec 2017
+# Johan Liakka
+# Yanchun He, NERSC
+# Last update Oct 2018
 
 # STRATEGY: Split the data into chucks of 10 years each.
 # Run the 10 chunk year in parallel for 3D operations
@@ -90,11 +91,26 @@ do
             let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
             yr_prnt=`printf "%04d" ${YR}`
             filename=${casename}.micom.hbgcy.${yr_prnt}.nc
-            eval $NCKS -O -v $var_list --no_tmp_fl $pathdat/$filename $WKDIR/${casename}_ANN_${yr_prnt}.nc &
-            pid+=($!)
+
+            fflag=1     #check if all required annual ts files exist
+            for var in $(echo $var_list | sed 's/,/ /g'|sed 's/pddpo//') ; do
+                tsfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f  $tsdir/ann_ts/${tsfile} ]; then
+                    echo "$tsdir/ann_ts/${tsfile} does not exist, ncks required variables..."
+                    fflag=0
+                    break
+                fi
+            done
+            if [ $fflag  = 0 ]; then
+                eval $NCKS -O -v $var_list --no_tmp_fl $pathdat/$filename $WKDIR/${casename}_ANN_${yr_prnt}.nc &
+                pid+=($!)
+            else
+                echo Skip computing year ${yr_prnt}, time-series variables already exist.
+            fi
+
             let iproc++
         done
-        for ((m=0;m<=$nyrsm;m++))
+        for ((m=0;m<${#pid[*]};m++))
         do
             wait ${pid[$m]}
             if [ $? -ne 0 ]; then
@@ -119,11 +135,24 @@ do
                 filename=${casename}.micom.hbgcm.${yr_prnt}-${mon}.nc
                 filenames+=($filename)
             done
-            eval $NCRA -O --no_tmp_fl --hdr_pad=10000 -w 31,28,31,30,31,30,31,31,30,31,30,31 -v $var_list -p $pathdat ${filenames[*]} $WKDIR/${casename}_ANN_${yr_prnt}.nc &
-            pid+=($!)
+            fflag=1     #check if all required annual ts files exist
+            for var in $(echo $var_list | sed 's/,/ /g'|sed 's/pddpo//') ; do
+                tsfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f  $tsdir/ann_ts/${tsfile} ]; then
+                    echo "$tsdir/ann_ts/${tsfile} does not exist"
+                    fflag=0
+                    break
+                fi
+            done
+            if [ $fflag  = 0 ]; then
+                eval $NCRA -O --no_tmp_fl --hdr_pad=10000 -w 31,28,31,30,31,30,31,31,30,31,30,31 -v $var_list -p $pathdat ${filenames[*]} $WKDIR/${casename}_ANN_${yr_prnt}.nc &
+                pid+=($!)
+            else
+                echo Skip computing year ${yr_prnt}, time-series variables already exist.
+            fi
             let iproc++
         done
-               for ((m=0;m<=$nyrsm;m++))
+               for ((m=0;m<${#pid[*]};m++))
         do
             wait ${pid[$m]}
             if [ $? -ne 0 ]; then
@@ -142,7 +171,9 @@ do
         let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
         yr_prnt=`printf "%04d" ${YR}`
         filename=${casename}_ANN_${yr_prnt}.nc
-        $NCKS --quiet -d depth,0 -d x,0 -d y,0 -v parea $WKDIR/$filename >/dev/null 2>&1
+        if [ -f $WKDIR/$filename ]; then
+            $NCKS --quiet -d depth,0 -d x,0 -d y,0 -v parea $WKDIR/$filename >/dev/null 2>&1
+        fi
         if [ $? -ne 0 ]; then
             $NCKS -A -v parea -o $WKDIR/$filename $grid_file
         fi
@@ -151,7 +182,7 @@ do
     # Loop over variables and do some averaging...
     for var in `echo $var_list | sed 's/,/ /g'`
     do
-        # Mass weighted 3D averaging of temp and saln
+        # Mass weighted 3D averaging of nutrients
         if [ $var == o2 ] || [ $var == si ] || [ $var == po4 ] || \
            [ $var == no3 ] || [ $var == dissic ]; then
             echo "Mass weighted global average of $var (yrs ${YR_start}-${YR_end})"
@@ -162,12 +193,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                eval $NCWA --no_tmp_fl -O -v $var -w pddpo -a sigma,y,x $WKDIR/$infile $WKDIR/$outfile &
-                pid+=($!)
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    eval $NCWA --no_tmp_fl -O -v $var -w pddpo -a sigma,y,x $WKDIR/$infile $WKDIR/$outfile &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
@@ -187,12 +220,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                     outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'epc100_area=epc100*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'epc100_tot=epc100_area.total($x,$y)*12.011*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v epc100_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'epc100_area=epc100*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'epc100_tot=epc100_area.total($x,$y)*12.011*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v epc100_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -205,12 +240,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                     outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'epcalc100_area=epcalc100*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'epcalc100_tot=epcalc100_area.total($x,$y)*12.0*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v epcalc100_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'epcalc100_area=epcalc100*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'epcalc100_tot=epcalc100_area.total($x,$y)*12.0*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v epcalc100_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -223,12 +260,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                     outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'co2fxd_area=co2fxd*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'co2fxd_tot=co2fxd_area.total($x,$y)*86400.0*365.0*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v co2fxd_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'co2fxd_area=co2fxd*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'co2fxd_tot=co2fxd_area.total($x,$y)*86400.0*365.0*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v co2fxd_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -241,12 +280,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'co2fxu_area=co2fxu*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'co2fxu_tot=co2fxu_area.total($x,$y)*86400.0*365.0*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v co2fxu_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'co2fxu_area=co2fxu*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'co2fxu_tot=co2fxu_area.total($x,$y)*86400.0*365.0*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v co2fxu_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -259,12 +300,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'dmsflux_area=dmsflux*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'dmsflux_tot=dmsflux_area.total($x,$y)*86400.0*365.0*62.13*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v dmsflux_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'dmsflux_area=dmsflux*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'dmsflux_tot=dmsflux_area.total($x,$y)*86400.0*365.0*62.13*1.0e-12' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v dmsflux_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -278,12 +321,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                eval $NCAP2 -O -s 'pp_vol=pp*parea*pddpo' $WKDIR/$infile $WKDIR/$outfile_tmp &
-                pid+=($!)
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    eval $NCAP2 -O -s 'pp_vol=pp*parea*pddpo' $WKDIR/$infile $WKDIR/$outfile_tmp &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
@@ -299,12 +344,14 @@ do
             do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                $NCAP2 -O -s 'pp_tot=pp_vol.total($sigma,$y,$x)*12.0*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp &
-                pid+=($!)
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'pp_tot=pp_vol.total($sigma,$y,$x)*12.0*86400.0*365.0*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
@@ -320,13 +367,15 @@ do
             do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                eval $NCKS --no_tmp_fl -O -v pp_tot $WKDIR/$outfile_tmp $WKDIR/$outfile &
-                pid+=($!)
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    eval $NCKS --no_tmp_fl -O -v pp_tot $WKDIR/$outfile_tmp $WKDIR/$outfile &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
@@ -356,12 +405,14 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCAP2 -O -s 'ppint_area=ppint*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'ppint_tot=ppint_area.total($x,$y)*86400.0*365.0*12*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v ppint_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ -f $WKDIR/$infile ] && [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCAP2 -O -s 'ppint_area=ppint*parea' $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'ppint_tot=ppint_area.total($x,$y)*86400.0*365.0*12*1.0e-15' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v ppint_tot $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -373,26 +424,35 @@ do
         let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
         yr_prnt=`printf "%04d" ${YR}`
         filename=${casename}_ANN_${yr_prnt}.nc
-        rm -f $WKDIR/$filename
+        if [ -f  $WKDIR/$filename ]; then
+            rm -f $WKDIR/$filename
+        fi
         let iproc++
     done
     let ichunk++
 done
 
 # Concancate files
+if [ ! -d $tsdir/ann_ts ]; then
+    mkdir -p $tsdir/ann_ts
+fi
+let "nyrs = $last_yr - $first_yr + 1"
 first_var=1
 for var in `echo $var_list | sed 's/,/ /g'`
 do
-    first_file=${var}_${casename}_ANN_${first_yr_prnt}.nc
-    if [ -f $WKDIR/$first_file ]; then
+    mv $WKDIR/${var}_${casename}_ANN_${filetype}_*.nc $tsdir/ann_ts/ >/dev/null 2>&1
+    first_file=${var}_${casename}_ANN_${filetype}_${first_yr_prnt}.nc
+    if [ -f $tsdir/ann_ts/$first_file ]; then
         echo "Merging all $var time series files..."
-        $NCRCAT -3 --no_tmp_fl -O $WKDIR/${var}_${casename}_ANN_????.nc $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc
+        #$NCRCAT -3 --no_tmp_fl -O $WKDIR/${var}_${casename}_ANN_????.nc $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc
+        $NCRCAT -3 --no_tmp_fl -O -p $tsdir/ann_ts -n ${nyrs},4,1 ${first_file} -o $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
         if [ $? -eq 0 ]; then
             if [ $first_var -eq 1 ]; then
                 first_var=0
-                mv $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc $tsdir/$ann_ts_file
+                mv $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc $tsdir/$ann_ts_file
             else
-                $NCKS -A -o $tsdir/$ann_ts_file $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc
+                $NCKS -A -o $tsdir/$ann_ts_file $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
+                rm -f $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
             fi
         else
             echo "ERROR: $NCRCAT -3 --no_tmp_fl -O $WKDIR/${var}_${casename}_ANN_????.nc $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc"
